requirements.txt
Django==4.1.3
python-dotenv==1.0.0
ShopifyAPI==12.3.0
requests==2.28.1